% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStanPred.R
\name{addStanPred}
\alias{addStanPred}
\title{Adds predictions and quantiles to data from Stan model}
\usage{
addStanPred(model, data, sim, q1 = 0.1, q2 = 0.9)
}
\arguments{
\item{model}{a Stan fit obtained with rstan::stan()}

\item{data}{dataframe with data used by the Stan model}

\item{sim}{character with simulated data from Stan model in Generated Quantities}

\item{q1}{low value for quantiles}

\item{q2}{high value for quantiles}
}
\value{
adds 3 columns to data : pred, q10 and q90
}
\description{
Adds predictions and quantiles to data from Stan model
}
