% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armaTimeSeries.R
\name{armaTimeSeries}
\alias{armaTimeSeries}
\title{generates arma time series with shifts}
\usage{
armaTimeSeries(N, p, q, e = 0.1)
}
\arguments{
\item{N}{number of time points to generate}

\item{p}{order of auto regressive model}

\item{q}{order of auto moving average model}

\item{e}{sd of error in model}
}
\value{
vector of N numeric
}
\description{
auto regressive moving average time series generation - based on
https://stats.stackexchange.com/questions/184828/create-random-time-series-with-shifts-in-r
with optional shifts in sequence
}
\examples{
armaTimeSeries (N = 100, p = 5, q = 20, e = 1)

}
