% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStanParam.R
\name{getStanParam}
\alias{getStanParam}
\title{returns data_fame with mean and quantiles of Stan model parameters}
\usage{
getStanParam(model, param.name)
}
\arguments{
\item{model}{a Stan fit obtained with rstan::stan()}

\item{param.name}{name (char) of parameter to get}
}
\value{
dataframe with 4 columns : parameter names, mean, 10 and 90% quantiles
}
\description{
returns data_fame with mean and quantiles of Stan model parameters
}
