% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize numeric vector}
\usage{
standardize(x, log = F)
}
\arguments{
\item{x}{numeric vector}

\item{log}{if True log of x is standardized}
}
\value{
numeric with standardized values s
}
\description{
returns numeric vector with standarzized values
}
