% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeDT.R
\name{standardizeDT}
\alias{standardizeDT}
\title{Standardize columns of data table}
\usage{
standardizeDT(dt, cols = "all.numeric", log = F)
}
\arguments{
\item{dt}{data frame of data table}

\item{cols}{default processes all numeric cols, otherwise specified column names}

\item{log}{if True returns  standardized log}
}
\value{
dt with additional columns with '_std' added to name
}
\description{
returns DT with additional standardized numeric columns with '_std' appended to name
}
