% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStanParam.R
\name{plotStanParam}
\alias{plotStanParam}
\title{Returns ggplot with pointrange for parameters}
\usage{
plotStanParam(model, param.name)
}
\arguments{
\item{model}{a Stan fit obtained with rstan::stan()}

\item{param.name}{name (char) of parameter to get}
}
\value{
ggplot with point range for each parameter
}
\description{
Returns ggplot with pointrange for each parameter of Stan model
}
