% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPotData.R
\name{simPotData}
\alias{simPotData}
\title{Simulate pot data for test}
\usage{
simPotData(
  N = 10,
  np = c(2, 3, 5),
  Y = c(-1, 0, 1),
  sigma = c(1, 1, 1),
  sigma_g = c(0.1, 0.1, 0.1)
)
}
\arguments{
\item{N}{number of days in simulation}

\item{np}{vector of number of pots per group}

\item{Y}{vector with mean effect for each group}

\item{sigma}{vector of sd for effect for pots}

\item{sigma_g}{vector of sd of pot means within groups}
}
\value{
dt with columns dat, C as pot , G as group, B as block type and Y effect
}
\description{
Create 'fake' pot data to test models
}
